refactor & bug fixes

+++ processor/src/dram.cpp
@@ -1,11 +1,12 @@
 #include <cstring>
+#include <iostream>
 #include "dram.hpp"
 
 uint64_t processor::dram::load(uint64_t addr, uint8_t bytes) const {
     uint64_t data = 0;
 
     for (int i = 0, sh = 0; i < bytes; i++, sh += 8) {
-        data |= mem[addr + i] << sh;
+        data |= (uint64_t) mem[addr + i] << sh;
     }
 
     return data;
